requirements.txt
absl-py==2.0.0
appdirs==1.4.4
asttokens==2.4.1
backcall==0.2.0
blosc2==2.0.0
cachetools==5.3.2
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
clean-fid==0.1.23
click==8.1.7
cmake==3.27.9
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.6
decorator #==5.1.1
deepdish==0.3.7
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.24.0
google-auth-oauthlib==1.0.0
grpcio==1.59.3
h5py==3.10.0
idna==2.10
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
iopath==0.1.10
ipdb
ipykernel==6.25.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==17.0.6
lpips==0.1.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.1
numexpr==2.8.6
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==10.1.0
plotly==5.18.0
portalocker==2.8.2
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
# pytorch-msssim @ git+https://github.com/LemonATsu/pytorch-msssim.git@f77a2446d0688f88ab935ea2e0fdada7691b83b0
# pytorch3d==0.3.0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==1.38.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
smplx==0.1.28
stack-data==0.6.3
sympy==1.12
tables==3.8.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tifffile==2023.7.10
torch==2.0.0
torchvision #==0.16.1
tqdm==4.66.1
traitlets==5.13.0
triton==2.0.0
typing_extensions==4.8.0
urllib3==1.26.18
wandb==0.16.1
wcwidth==0.2.12
Werkzeug==3.0.1
yacs==0.1.8
zipp==3.17.0
pytorch-msssim

